+++ package.json
@@ -15,13 +15,13 @@
 	"dependencies": {
 		"@chakra-ui/react": "^2.7.1",
 		"@clerk/clerk-react": "^4.22.1",
+		"@crossbell/connect-kit": "1.5.8",
 		"@crossbell/indexer": "^1.5.2",
 		"@crossbell/notification": "^1.5.0",
 		"@crossbell/ui": "^1.5.0",
 		"@emotion/react": "^11.11.1",
 		"@emotion/styled": "^11.11.0",
 		"@ethersproject/providers": "^5.7.2",
-		"@flarezone/connect-kit": "^1.4.9",
 		"@mantine/core": "^6.0.17",
 		"@radix-ui/react-dialog": "^1.0.4",
 		"@radix-ui/react-icons": "^1.3.0",
